refactor: move bash to home-manager
My preference is to use darwin whenever possible to decrease dependencies. But it seems like I need home manager for bash so that I can use config like historyControl and historyFileSize, which are not available in nix darwin.

I'm also copying my bashrc as is. Still need to work on making it more generic.

Signed-off-by: Samer Almasri <almasri@ualberta.ca>