chore: withdraw control menu bar todo
I can't seem to find a way to do this. I tried to run `defaults read` twice, before and after the changing the settings that show the sound on the menu bar. Comparing the results suggests that mac now uses an event store to manage this. I get a change in this property: `DKKnowledgeStorageDidInsertEventsNotification`.

I also tried `defaults write com.apple.controlcenter "NSStatusItem Visible Bluetooth" -bool true` (to enable bluetooth) with no luck

Signed-off-by: Samer Almasri <almasri@ualberta.ca>